+++ src/heat.js
@@ -74,6 +74,8 @@
                     bindingOptions = buildAttributeOptions( bindingOptions.result );
                     bindingOptions.currentView = {};
                     bindingOptions.currentView.element = element;
+                    bindingOptions.currentView.mapContents = null;
+                    bindingOptions.currentView.mapContentsScrollLeft = 0;
                     bindingOptions.currentView.colorsVisible = {};
                     bindingOptions.currentView.year = bindingOptions.year;
                     bindingOptions.currentView.type = _elements_DateCounts_DefaultType;
@@ -109,6 +111,10 @@
     }
 
     function renderControl( bindingOptions ) {
+        if ( isDefined( bindingOptions.currentView.mapContents ) ) {
+            bindingOptions.currentView.mapContentsScrollLeft = bindingOptions.currentView.mapContents.scrollLeft;
+        }
+
         bindingOptions.currentView.element.className = "heat-js";
         bindingOptions.currentView.element.innerHTML = _string.empty;
 
@@ -167,13 +173,13 @@
     }
 
     function renderControlMap( bindingOptions ) {
-        var mapContents = createElement( bindingOptions.currentView.element, "div", "map-contents" ),
-            map = createElement( mapContents, "div", "map" );
+        bindingOptions.currentView.mapContents = createElement( bindingOptions.currentView.element, "div", "map-contents" );
+
+        var map = createElement( bindingOptions.currentView.mapContents, "div", "map" ),
+            currentYear = bindingOptions.currentView.year,
+            monthAdded = false;
 
         renderControlViewGuide( bindingOptions );
-
-        var currentYear = bindingOptions.currentView.year,
-            monthAdded = false;
 
         if ( bindingOptions.showDayNames ) {
             var days = createElement( map, "div", "days" );
@@ -258,6 +264,10 @@
 
                 monthAdded = true;
             }
+        }
+        
+        if ( bindingOptions.keepScrollPositions ) {
+            bindingOptions.currentView.mapContents.scrollLeft = bindingOptions.currentView.mapContentsScrollLeft;
         }
     }
 
@@ -553,6 +563,7 @@
         options.exportOnlyYearBeingViewed = getDefaultBoolean( options.exportOnlyYearBeingViewed, true );
         options.year = getDefaultNumber( options.year, new Date().getFullYear() );
         options.showDayNumbers = getDefaultBoolean( options.showDayNumbers, false );
+        options.keepScrollPositions = getDefaultBoolean( options.keepScrollPositions, false );
 
         if ( isInvalidOptionArray( options.monthsToShow ) ) {
             options.monthsToShow = [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12 ];